requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
decorator==5.1.1
et-xmlfile==1.1.0
execnet==1.9.0
idna==3.3
iniconfig==1.1.1
jsonpath==0.82
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
openpyxl==3.0.9
packaging==21.3
pbr==5.8.1
pluggy==1.0.0
ply==3.11
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
pytest-assume==2.4.3
pytest-forked==1.4.0
pytest-xdist==2.5.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
